DataExtraction/requirements.txt
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
flake8==7.1.1
Flask==2.0.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
mccabe==0.7.0
mysql-connector-python==8.0.27
protobuf==3.19.1
pycodestyle==2.12.1
pyflakes==3.2.0
requests==2.26.0
soupsieve==2.3
urllib3==1.26.7
Werkzeug==2.0.2
